Excel files that describe specs
The Wiring file shows how the LCD and rotary encoders are wired to the Raspberry Pi. The pin numbers are configured in the Pyton file erres_variables.py. The Interface excel file is used to experiment with layouts on the 4x20 LCD, and to create new characters and retrieve their code.